PDJB-1131: Restructure files and names to be a little more sensible
