+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/journeys/landlordRegistration/LandlordRegistrationJourneyFactory.kt
@@ -28,8 +28,8 @@ import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.
 import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.PrivacyNoticeStep
 import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.tasks.IdentityTask
 import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.tasks.LandlordRegistrationAddressTask
-import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.tasks.LandlordRegistrationNotOrgLandlordTask
-import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.tasks.LandlordRegistrationTask
+import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.tasks.LandlordRegistrationForNotOrgLandlordTask
+import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.tasks.LandlordRegistrationOldTask
 import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.FinishCyaJourneyStep
 import uk.gov.communities.prsdb.webapp.journeys.shared.stepConfig.LookupAddressStep
 import uk.gov.communities.prsdb.webapp.journeys.shared.stepConfig.ManualAddressStep
@@ -52,7 +52,7 @@ class LandlordRegistrationJourneyFactory(
         return if (checkingAnswersFor == null) {
             mainJourneyMap(state)
         } else {
-            LandlordRegistrationTask.checkYourAnswersJourneyMap(state, checkingAnswersFor)
+            LandlordRegistrationOldTask.checkYourAnswersJourneyMap(state, checkingAnswersFor)
         }
     }
 
@@ -63,12 +63,12 @@ class LandlordRegistrationJourneyFactory(
             configure {
                 withAdditionalContentProperty { "title" to "registerAsALandlord.title" }
             }
-            task(journey.landlordRegistrationTask) {
+            task(journey.landlordRegistrationOldTask) {
                 initialStep()
                 nextStep { journey.deleteJourneyStep }
             }
             step(journey.deleteJourneyStep) {
-                parents { journey.landlordRegistrationTask.isComplete() }
+                parents { journey.landlordRegistrationOldTask.isComplete() }
                 nextUrl { LANDLORD_REGISTRATION_CONFIRMATION_ROUTE }
             }
         }
@@ -79,8 +79,8 @@ class LandlordRegistrationJourneyFactory(
 @JourneyFrameworkComponent("landlordRegistrationJourney")
 class LandlordRegistrationJourney(
     // Landlord registration task
-    override val landlordRegistrationTask: LandlordRegistrationTask,
-    override val landlordRegistrationNotOrgLandlordTask: LandlordRegistrationNotOrgLandlordTask,
+    override val landlordRegistrationOldTask: LandlordRegistrationOldTask,
+    override val landlordRegistrationForNotOrgLandlordTask: LandlordRegistrationForNotOrgLandlordTask,
     // Privacy notice step
     override val privacyNoticeStep: PrivacyNoticeStep,
     // Identity task
