+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/journeys/landlordRegistration/tasks/LandlordRegistrationForNotOrgLandlordTask.kt
@@ -13,7 +13,7 @@ import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.
 import uk.gov.communities.prsdb.webapp.journeys.shared.stepConfig.AbstractCheckYourAnswersStep
 
 @JourneyFrameworkComponent
-class LandlordRegistrationNotOrgLandlordTask : Task<LandlordRegistrationState>() {
+class LandlordRegistrationForNotOrgLandlordTask : Task<LandlordRegistrationState>() {
     override fun makeSubJourney(state: LandlordRegistrationState) =
         subJourney(state) {
             step(journey.emailStep) {
